requirements.txt
telethon>=1.34.0
cryptg>=0.4.0
pysocks>=1.7.1
aiohttp>=3.9.0
psutil>=5.9.0
Pillow>=10.0.0
Flask>=3.0.0
SpeechRecognition>=3.10.0
soundfile>=0.12.1
cryptography>=41.0.0
qrcode[pil]>=7.4.0
emoji>=2.0.0
